Move app router to root index.js
