+++ src/components/NotificationCenter/components/Notification/index.js
