+++ src/utils/removeQueryString.js
